Removed 'SetIsInitialBlockDownload' from the IBD interface. Added mock for unit tests.
