+++ src/Stratis.Bitcoin.Features.Consensus/InitialBlockDownloadState.cs
@@ -15,11 +15,11 @@ namespace Stratis.Bitcoin.Features.Consensus
     /// <seealso cref="IInitialBlockDownloadState" />
     public class InitialBlockDownloadState : IInitialBlockDownloadState
     {
+        /// <summary>A provider of the date and time.</summary>
+        protected readonly IDateTimeProvider dateTimeProvider;
+
         /// <summary>Provider of block header hash checkpoints.</summary>
         private readonly ICheckpoints checkpoints;
-
-        /// <summary>A provider of the date and time.</summary>
-        private readonly IDateTimeProvider dateTimeProvider;
 
         /// <summary>Information about node's chain.</summary>
         private readonly ChainState chainState;
@@ -29,12 +29,6 @@ namespace Stratis.Bitcoin.Features.Consensus
 
         /// <summary>User defined node settings.</summary>
         private readonly NodeSettings nodeSettings;
-
-        /// <summary>Time until IBD state can be checked.</summary>
-        private DateTime lockIbdUntil;
-
-        /// <summary>A cached result of the IBD method.</summary>
-        private bool ibdCached;
 
         /// <summary>
         /// Creates a new instance of the <see cref="InitialBlockDownloadState" /> class.
@@ -50,16 +44,11 @@ namespace Stratis.Bitcoin.Features.Consensus
             this.chainState = chainState;
             this.checkpoints = checkpoints;
             this.dateTimeProvider = DateTimeProvider.Default;
-
-            this.lockIbdUntil = DateTime.MinValue;
         }
 
         /// <inheritdoc />
-        public bool IsInitialBlockDownload()
+        public virtual bool IsInitialBlockDownload()
         {
-            if (this.lockIbdUntil >= this.dateTimeProvider.GetUtcNow())
-                return this.ibdCached;
-
             if (this.chainState == null)
                 return false;
 
@@ -76,13 +65,6 @@ namespace Stratis.Bitcoin.Features.Consensus
                 return true;
 
             return false;
-        }
-
-        /// <inheritdoc />
-        public void SetIsInitialBlockDownload(bool blockDownloadState, DateTime lockStateUntil)
-        {
-            this.lockIbdUntil = lockStateUntil;
-            this.ibdCached = blockDownloadState;
         }
     }
 }
