+++ src/Stratis.Bitcoin.IntegrationTests/EnvironmentMockUpHelpers/NodeRunners.cs
@@ -1,8 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Text;
+using System.Linq;
 using NBitcoin;
 using NBitcoin.Protocol;
 using Stratis.Bitcoin.Builder;
@@ -14,6 +13,10 @@ using Stratis.Bitcoin.Features.Miner;
 using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Features.Wallet;
 using Stratis.Bitcoin.Utilities;
+using Microsoft.Extensions.DependencyInjection;
+using Stratis.Bitcoin.Builder.Feature;
+using Stratis.Bitcoin.IntegrationTests.EnvironmentMockUpHelpers;
+using Stratis.Bitcoin.Interfaces;
 
 namespace Stratis.Bitcoin.IntegrationTests
 {
@@ -103,6 +106,7 @@ namespace Stratis.Bitcoin.IntegrationTests
                     .UseWallet()
                     .AddPowPosMining()
                     .AddRPC()
+                    .MockIBD()
                     .Build();
             }
 
@@ -181,6 +185,7 @@ namespace Stratis.Bitcoin.IntegrationTests
                 .UseWallet()
                 .AddPowPosMining()
                 .AddRPC()
+                .MockIBD()
                 .Build();
 
             return fullNode;
@@ -241,6 +246,7 @@ namespace Stratis.Bitcoin.IntegrationTests
                     .AddMining()
                     .UseWallet()
                     .AddRPC()
+                    .MockIBD()
                     .Build();
             }
 
@@ -248,5 +254,28 @@ namespace Stratis.Bitcoin.IntegrationTests
         }
 
         public FullNode FullNode;
+    }
+
+    public static class FullNodeTestBuilderExtension
+    {
+        public static IFullNodeBuilder MockIBD(this IFullNodeBuilder fullNodeBuilder)
+        {
+            fullNodeBuilder.ConfigureFeature(features =>
+            {
+                foreach (IFeatureRegistration feature in features.FeatureRegistrations)
+                {
+                    feature.FeatureServices(services =>
+                    {
+                        // Get default IBD implementation and replace it with the mock.
+                        ServiceDescriptor ibdService = services.FirstOrDefault(x => x.ServiceType == typeof(IInitialBlockDownloadState));
+                        services.Remove(ibdService);
+
+                        services.AddSingleton<IInitialBlockDownloadState, MockedInitialBlockDownloadState>();
+                    });
+                }
+            });
+
+            return fullNodeBuilder;
+        }
     }
 }
