+++ src/Stratis.Bitcoin/Interfaces/IInitialBlockDownloadState.cs
@@ -14,13 +14,5 @@ namespace Stratis.Bitcoin.Interfaces
         /// </summary>
         /// <returns><c>true</c> if the node is currently doing IBD, <c>false</c> otherwise.</returns>
         bool IsInitialBlockDownload();
-
-        /// <summary>
-        /// Sets last IBD status update time and result.
-        /// <para>Used in tests only.</para>
-        /// </summary>
-        /// <param name="blockDownloadState">New value for the IBD status, <c>true</c> means the node is considered in IBD.</param>
-        /// <param name="lockStateUntil">Time until IBD state won't be changed.</param>
-        void SetIsInitialBlockDownload(bool blockDownloadState, DateTime lockStateUntil);
     }
 }
